Add workaround to fix Continuous
srcdir != builddir works only for tarball builds, not for the git builds
that Continuous is doing. The problem is that the generated vapi and
header file are expected to be in srcdir (where they would be for a
tarball build) rather than builddir (where they end up during a git
build). This is hard to fix.

https://bugzilla.gnome.org/show_bug.cgi?id=751624