scores: Tolerate some messiness in the CategoryConvertFunc
If the CategoryConvertFunc does not closely examine the values it
receives -- say, if it's a test function that always returns the same
category -- then bad things will happen when it is passed the "scores"
category, corresponding to the new directory created by this function.
There is no sensible thing a game could ever do with this besides to
return null, so just skip it.