Add API version to include directory and library name
This should have been done from the start; it was a serious oversight.
Unfortunately it means all dependent games (gnome-klotski, gnome-robots,
gnome-nibbles, gnome-mines) will need to be rebuilt. Fortunately no
changes in the games should be required, as the games all get the
include directory and library name from pkg-config.

Sneakily rename the include directory location while we're at it.

https://bugzilla.gnome.org/show_bug.cgi?id=766334