context: Make usable when created via g_object_new()
This is tricky. First, we need to add properties for all the constructor
parameters. But that doesn't work for the CategoryRequestFunc: the best
we can do there would be a void* property, which would be
non-introspectable, defeating the point of the whole exercise. Instead,
we can add a bool object to disable loading scores when the object is
created. Then a category request function can be set separately when
loading scores manually.

I'm not actually sure if that's introspectable either, TBH. Maybe it
doesn't matter, because we don't currently support introspection. But
hey, it's a step in the right direction.

Nobody will ever figure out how to do use this, because there's no
documentation. Would be good to add some gtkdoc....