scores: Use List instead of PriorityQueue to store scores
Instead, sort the scores right before they're needed. This should be
less efficient in theory, but our use of PriorityQueue is not working
properly and it's not clear why. Sorting on demand simplifies several
functions, and should be just fine.

This fixes scores appearing out of order on the scores dialog.

https://bugzilla.gnome.org/show_bug.cgi?id=789399