context: Fix username not being saved in high scores dialog
When this library was ported from GTK 3 to GTK 4, we switched from
actually running the dialog (and waiting for a response, indicating the
dialog has been closed) to just showing the dialog. But
add_score_internal() was not updated and still assumed the original
semantics of run_dialog(). Also, the name of this function strongly
implies the dialog should be "run" (in a nested loop), not just shown.

I think "running" is a bad idea. "Showing" is much safer. However,
for now let's restore the original behavior. Otherwise,
add_score_internal() performs the final save immediately after the
dialog is shown, rather than after it is closed, as the code assumes.

Fixes #24