Fix circular refs with delegate functions
The CategoryRequestFunc delegate keeps a ref on the Context object,
which is a problem because each Context owns its CategoryRequestFunc.
This makes it impossible for any Context to ever be freed.

DirectoryImporter has the same problem with its CategoryConvertFunc, and
HistoryFileImporter has the same problem with its HistoryConvertFunc.

In all cases, the delegates cannot be unowned: the object needs to hold
a ref on the delegate. And I'm not sure how to tell Vala that the
delegate does not need to ref the object, even though that is
unnecessary and would be the logical way to break this cycle in other
languages.

This design error might have been very hard to fix, but fortunately in
all cases the delegates are only required for a limited time, so we can
simply unset them when no longer needed.

Fixes #27