scores: add extra_info
extra_info lets games add custom data to scores, because
games might need add important context to a score that isn't
possible with what is currently in Score.
Storing extra_info turned out to be more complex than I thought.
We need to keep every score in one line and we also can't add
more spaces, because the values are separated by spaces. So
we encode the string in base 64. But we have to make this value
part of an existing column, so we seperate it from the score
value using '#'.
It is important to note that this change is pretty much permanent
and changes to the score file format can not be easily undone.

closes: #35